moved logo image to new assets folder ( #14 )
