+++ src/components/DashBoard.jsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { Layout, } from 'antd'
 import HomeEmpty from './HomeEmpty'
-import logo from './420logo.png';
+import logo from './assets/420logo.png';
 import './css/DashBoard.css';
 import { accountActions } from '../actions'
 import { connect } from 'react-redux'
