+++ src/components/LoginPage.jsx
@@ -3,7 +3,7 @@ import { Button, Col, Checkbox, Form, Icon, Input, Row, Layout } from 'antd';
 import './css/LoginPage.css'
 import { accountActions } from '../actions'
 import { connect } from 'react-redux'
-import logo from './420logo.png';
+import logo from './assets/420logo.png';
 
 class LoginForm extends React.Component {
     state = {
