+++ src/components/RegisterPage.jsx
@@ -3,7 +3,7 @@ import { Button, Col, Form, Icon, Input, Layout, Row } from 'antd';
 import './css/RegisterPage.css'
 import { accountActions } from '../actions'
 import { connect } from 'react-redux'
-import logo from './420logo.png';
+import logo from './assets/420logo.png';
 
 class RegisterForm extends React.Component {
